requirements.txt
pillow==10.0.0
pytesseract==0.3.10
python-docx==0.8.11
